# refactor(charts): slow single page to fast multiple tabs (solve: #4308)