### Description
- #4308 
- I refactored slow single page using `<Tabs />` component supported by shadcn/ui.
- The result is pretty remarkable. See the following benchmark results.
- **Total Blocking Time** : `2720ms -> 1020ms` **(62.5% Improved)**
- **Speed Index** : `2.7s -> 1.3s` **(51.8% Improved)**

### System
- Google Chrome | 126.0.6478.185 (arm64)
- OS | macOS Version 14.5(Build 23F79)
- JavaScript | V8 12.6.228.28
- User Agent | Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/126.0.0.0 Safari/537.36

### Additional
- The only difference is you can't share the link with selected tab. 
	- related issue: #414
- I'm willing to contribute! Let me know if it's worth doing so.

### As-Is
<img width="2032" alt="스크린샷 2024-07-29 오후 10 04 34" src="https://github.com/user-attachments/assets/699585a6-5706-49ee-acc5-621e21ef6695">

### To-Be
<img width="2032" alt="스크린샷 2024-07-29 오후 10 01 10" src="https://github.com/user-attachments/assets/65ff1fb5-8de4-46b6-bd97-24182fa151d7">
- The reason accessibility dropped from 100 to 96 is that the default color set of the <TabsTrigger /> is not very accessibility-friendly.
<img width="627" alt="image" src="https://github.com/user-attachments/assets/68f9d118-b6f2-431e-82f5-3f5359484be6">