@devymarry 2024-08-02
> The only difference is you can't share the link with selected tab.

I agree. The tab component from Shadcn doesn’t support preserving the active tab, so using query parameters might be a great solution.