@shadcn 2024-08-05
Woah! This is something I planned to do. Thanks for starting the work on this. I'll review and merge your PR (might push a few updates to it)