@Henry-Hong 2024-08-05
Thanks! I’m excited to contribute to this project—I’m a big fan of it. If there’s anything else I can help with, just let me know.

I’m happy to put together an additional PR to handle **preserving the active tab using query parameters** if that would be useful. Just let me know if you’d like me to take that on.

Thanks again!